Remove `FilteringLinkExtractor`
